+++ src/vscode/configuration/PowerBIConfiguration.ts
@@ -95,6 +95,10 @@ export abstract class PowerBIConfiguration {
 	static get clientId(): string { return this.getValue("clientId"); }
 	static set clientId(value: string) { this.setValue("clientId", value); }
 
+	static get workspaceFilter(): string { return this.getValue("workspaceFilter"); }
+	static set workspaceFilter(value: string) { this.setValue("workspaceFilter", value); }
+	static get workspaceFilterRegEx(): RegExp { return new RegExp(this.getValue("workspaceFilter")); }
+
 	static get tmdlClientId(): string { return this.getValue("TMDL.clientId"); }
 	static set tmdlClientId(value: string) { this.setValue("TMDL.clientId", value); }