+++ src/ThisExtension.ts
@@ -180,7 +180,7 @@ export abstract class ThisExtension {
 
 	static async initializeLogger(context: vscode.ExtensionContext): Promise<void> {
 		if (!this._logger) {
-			this._logger = vscode.window.createOutputChannel("PowerBI.VSCode");
+			this._logger = vscode.window.createOutputChannel("Power BI Studio");
 			this.log("Logger initialized!");
 
 			context.subscriptions.push(this._logger);